+++ CHANGELOG.md
@@ -5,7 +5,8 @@ XXXX-XX-XX, v0.6.11
 -------------------
 
 - closed [#132](https://github.com/n1k0/casperjs/issues/132) - added ability to include js/coffee files using a dedicated option when using the `casper test` command
-- fixes [#140](https://github.com/n1k0/casperjs/issues/140) - `casper test` now resolves local paths urls
+- fixed [#140](https://github.com/n1k0/casperjs/issues/140) - `casper test` now resolves local paths urls
+- fixed [#148](https://github.com/n1k0/casperjs/issues/148) - `utils.isWebPage()` was broken
 - closed [#144](https://github.com/n1k0/casperjs/issues/144) - added a `safeLogs` option to blur password values in debug logs. **This option is set to `true` by default.**
 - added [`Casper.userAgent()`](http://casperjs.org/api.html#casper.userAgent) to ease a more dynamic setting of user-agent string
 - added `Tester.assertTitleMatch()` method